requirements.txt
openai==1.2.3
fastapi
llama-cpp-python
